fixed bug with new parameter

+++ general/app/bots/binance_bot.py
@@ -163,7 +163,7 @@ async def creating_orders_bot(session, redis, shared_data, client, stop_event):
         timeoutOccurred = False
 
         try:
-            timeout = bot_config.time_to_wait_for_entry_price_to_open_order_in_minutes * 60
+            timeout = int(bot_config.time_to_wait_for_entry_price_to_open_order_in_minutes * 60)
 
             trade_type, entry_price = await asyncio.wait_for(
                 _wait_for_entry_price(
